[ci] Split up integration.shard record_use_test.dart
